// src/app/models/LocaisAcessoTaxas.ts
export class LocaisAcessoTaxas {
    localAcessoValorId: number;
    localAcessoId: number; 
    valor: number; 
    dataInicio: Date;
    dataInclusao: Date;
    dataAlteracao: Date;
    usuarioIdInclusao: number; 
    usuarioIdAlteracao: number; 
    dataInativacao: Date;
    usuarioIdInativacao: number; 
    ativo: boolean;
    usuarioResponsavel: string;
    dataInicioFormatado: string;
    valorTaxaFormatado: string;
}